NEW : Added SudokuFunctionFactory but not effective
We commented the parts where we define ActionsFunction and ResultFunction : as we can see, it needs a specific name for every possible action.
This does not seem possible with our problem right now since we have so much different possible actions, depending on location and number to fill.
The two actions right now we defined in SudokuBoard are FILL and ERASE, where we left comments wondering about their use.